Add new command to comms system
New command called "setAzimuth" allows changing the
mission["launchAzimuth"] in the passive guidance phase. This lets you
adjust the course of the rocket to pass over specific position on the
surface (useful for landing the first stage on a ship located somewhere
on the flight path). NOTE: This is only meant for small adjustments,
large changes will increase deltaV usage and might make it impossible
to reach orbit on some occasions.